+++ trace/provider.go
@@ -99,7 +99,7 @@ func NewProvider(opts ...Option) (Provider, error) {
 	spanProcesor := spanProcessorFactory(provider.cfg.SpanProcessorType, exporter)
 
 	// create the sampler based on the configs
-	sampler := getSampler(provider.cfg.SamplingType, provider.cfg.SamplingRate, provider.cfg.ParentBasedSampling)
+	sampler := getSampler(provider.cfg.Sampling.SamplerType, provider.cfg.Sampling.SamplingRate, provider.cfg.Sampling.ParentBasedSampling)
 
 	// Create the tracer provider
 	// The tracer provider will use the resource and exporter created previously
@@ -138,33 +138,6 @@ func NewProvider(opts ...Option) (Provider, error) {
 	provider.logger.Info("Tracer provider initialized successfully")
 
 	return provider, nil
-}
-
-func getSampler(samplingType string, samplingRate float64, parentBased bool) sdktrace.Sampler {
-	switch samplingType {
-	case config.ALWAYSON:
-		if parentBased {
-			return sdktrace.ParentBased(sdktrace.AlwaysSample())
-		} else {
-			return sdktrace.AlwaysSample()
-		}
-	case config.ALWAYSOFF:
-		if parentBased {
-			return sdktrace.ParentBased(sdktrace.NeverSample())
-		} else {
-			return sdktrace.NeverSample()
-		}
-	case config.TRACEIDRATIOBASED:
-		fmt.Println("will return a trace ID ratio sampler with sampling rate:", samplingRate)
-		if parentBased {
-			return sdktrace.ParentBased(sdktrace.TraceIDRatioBased(samplingRate))
-		} else {
-			return sdktrace.TraceIDRatioBased(samplingRate)
-		}
-	default:
-		// Default to AlwaysOn if no valid sampling type is provided
-		return sdktrace.AlwaysSample()
-	}
 }
 
 func (tp *traceProvider) Shutdown(ctx context.Context) error {
